user-service/requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
redis
pydantic
